Updated BPU organization | Added scaladoc comments
