+++ src/main/scala/bpu/misc/2bc-table.scala
